server/requirements.txt
flask==3.0.0
flask-cors==4.0.0
psycopg2-binary==2.9.9
pandas==2.1.4
numpy==1.26.2
scikit-learn==1.3.2
imbalanced-learn==0.11.0
lightgbm==4.1.0
xgboost==2.0.2
catboost==1.2.2
river==0.20.0